requirements.txt
python-binance 
numpy
pandas 
websocket_client
nsepy
ta
sklearn
#TA-Lib